requirements.txt
alembic==1.8.1
aniso8601==9.0.1
click==8.1.3
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Restless-NG==2.3.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
uuid==1.30
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.1
